Fix using a subquery in FROM with multiple GROUP BY
The issue is that `->from()` is trying to split and escape the tables sent to it, it's not expecting an entire SQL query.  It is having trouble escaping this and gives back a bad query.

The "fix" is to convert "GROUP BY a, b" into "GROUP BY (a), (b)".  The parenthesis tell CodeIgniter to stop trying to escape it (after being split on commas).

This is pretty ugly and I know I can make it better, but let's see if it works.